no beforeEach needed any longer
added declarations and schemas to testBed. I dont like it yet but for now it works. Added Try catch block to the logic that finds method on the prototypes of an objects because I had issues with some getters that tried to find things that are not there yet. e.g. deps of a service. magic.happens doesn't need a beforeEach any longer.